1493: Check no data missing before closing the case
